+++ _tests_/react-ts-tw/eslint.config.js
@@ -190,6 +190,7 @@ export default [ // eslint-disable-line
             "code-style/hook-callback-format": "error",
             "code-style/hook-deps-per-line": "error",
             "code-style/hook-file-naming-convention": "error",
+            "code-style/hook-function-naming-convention": "error",
             "code-style/if-else-spacing": "error",
             "code-style/if-statement-format": "error",
             "code-style/import-format": "error",